requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
